kma-api/requirements.txt
fastapi
uvicorn[standard]
requests
redis>=5.0.0
psycopg2-binary
